+++ lib/vector-core/buffers/Cargo.toml
@@ -23,7 +23,7 @@ pin-project = { version = "1.0.8", default-features = false }
 rkyv = { version = "0.7.25", default-features = false, features = ["size_32", "std", "strict", "validation"] }
 serde = { version = "1.0.130", default-features = false, features = ["derive"] }
 snafu = { version = "0.6.10", default-features = false, features = ["std"] }
-tokio = { version = "1.13.0", default-features = false, features = ["rt", "macros", "rt-multi-thread", "sync", "fs", "io-util", "time"] }
+tokio = { version = "1.13.0", default-features = false, features = ["rt", "macros", "rt-multi-thread", "sync", "fs", "io-util", "time", "tracing"] }
 tokio-stream = { version = "0.1", default-features = false, features = ["sync"] }
 tokio-util = { version = "0.6", default-features = false }
 tracing = { version = "0.1.29", default-features = false }
@@ -47,9 +47,7 @@ tracing = { version = "0.1.29", features = ["attributes"] }
 tracing-fluent-assertions = "0.1.3"
 tracing-subscriber = { version = "0.2.25" }
 human_bytes = "0.3.0"
-
-[features]
-helpers = []
+console-subscriber = { git = "https://github.com/tokio-rs/console" }
 
 [[bench]]
 name = "on_disk"
@@ -64,13 +62,4 @@ name = "in_memory_v2"
 harness = false
 
 [[example]]
-name = "disk_v1"
-required-features = ["helpers"]
-
-[[example]]
-name = "disk_v2"
-required-features = ["helpers"]
-
-[[example]]
-name = "in_memory_v2"
-required-features = ["helpers"]
+name = "buffer_perf"
